fault tolerance merged into main
